requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
beautifulsoup4
faiss-cpu
streamlit